+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Print a note with the list of alternative targets when the solver cannot find a suitable component version for the current target but there are some candidates for other targets.
 - Add the `examples` field for the custom examples paths
 - Add new environment variables `IDF_COMPONENT_REGISTRY_URL` and `IDF_COMPONENT_REGISTRY_PROFILE` for configuring the component manager
+- Exclude build artefacts from the examples
 
 ### Fixed